samples/requirements.txt
langchain
python-dotenv
langchain_community
langchain-text-splitters
langchain-openai
pypdf
faiss-cpu
